Add View and Controller classes;
The View class takes a Vedeu::Menu instance which has been constructed from the TrackCollection in the Controller. It passes this 'playlist' to the Vedeu::Parser which forwards it to the 'playlist' interface for rendering.

The Controller class orchestrates the firing of events on key presses by the user, invoking the Player etc.

An interface and method to start the application has also been added to the Playa::Application class.